+++ src/Program/Program.hpp
@@ -22,7 +22,7 @@ public:
 
   /* Program Component generation */
   void generatePrism(float, float, float, shared_ptr<Point> center);
-  void generateTexturedPrism(float, float, float, shared_ptr<Point>, GLuint, int);
+  void generateTexturedPrism(float, float, float, shared_ptr<Point>, GLuint, float);
   void generateLine(shared_ptr<Point> start, shared_ptr<Point> end);
   void generateGrid(float maxItem, float separator);
